Move debug logic into extension on ExposureManager
