+++ ios/BT/ExposureManager.swift
@@ -8,7 +8,7 @@ import BackgroundTasks
 final class ExposureManager: NSObject {
   
   @objc static let shared = ExposureManager()
-  
+
   private static let backgroundTaskIdentifier = Bundle.main.bundleIdentifier! + ".exposure-notification"
   
   let manager = ENManager()
@@ -39,7 +39,7 @@ final class ExposureManager: NSObject {
         }
       }
     }
-    
+
     // Schedule background task if needed whenever EN authorization status changes
     NotificationCenter.default.addObserver(self, selector: #selector(scheduleBackgroundTaskIfNeeded), name: .AuthorizationStatusDidChange, object: nil)
   }
@@ -133,6 +133,7 @@ final class ExposureManager: NSObject {
           finish(.failure(error))
           return
         }
+
         dispatchGroup.notify(queue: .main) {
           for result in localURLResults {
             switch result {
@@ -173,13 +174,38 @@ final class ExposureManager: NSObject {
     
     return progress
   }
+
+  @objc func registerBackgroundTask() {
+    notifyUserBlueToothOffIfNeeded()
+    BGTaskScheduler.shared.register(forTaskWithIdentifier: ExposureManager.backgroundTaskIdentifier, using: .main) { [weak self] task in
+
+      // Notify the user if bluetooth is off
+      self?.notifyUserBlueToothOffIfNeeded()
+
+      // Perform the exposure detection
+      let progress = ExposureManager.shared.detectExposures { success in
+        task.setTaskCompleted(success: success)
+      }
+
+      // Handle running out of time
+      task.expirationHandler = {
+        progress.cancel()
+        BTSecureStorage.shared.exposureDetectionErrorLocalizedDescription = NSLocalizedString("BACKGROUND_TIMEOUT", comment: "Error")
+      }
+
+      // Schedule the next background task
+      self?.scheduleBackgroundTaskIfNeeded()
+    }
+  }
   
   func getAndPostDiagnosisKeys(completion: @escaping (Error?) -> Void) {
     manager.getDiagnosisKeys { temporaryExposureKeys, error in
       if let error = error {
         completion(error)
       } else {
-        APIClient.shared.request(DiagnosisKeyListRequest.post((temporaryExposureKeys ?? []).compactMap { $0.asCodableKey }, [.US]), requestType: .postKeys) { result in
+        APIClient.shared.request(DiagnosisKeyListRequest.post((temporaryExposureKeys ?? []).compactMap { $0.asCodableKey },
+                                                              [.US]),
+                                 requestType: .postKeys) { result in
           switch result {
           case .success:
             break
@@ -190,106 +216,11 @@ final class ExposureManager: NSObject {
       }
     }
   }
-  
-  // Includes today's key, requires com.apple.developer.exposure-notification-test entitlement
-  func getAndPostTestDiagnosisKeys(completion: @escaping (Error?) -> Void) {
-    manager.getTestDiagnosisKeys { temporaryExposureKeys, error in
-      if let error = error {
-        completion(error)
-      } else {
-        APIClient.shared.request(DiagnosisKeyListRequest.post((temporaryExposureKeys ?? []).compactMap { $0.asCodableKey }, [.US]), requestType: .postKeys) { result in
-          switch result {
-          case .success:
-            completion(nil)
-          case .failure(let error):
-            completion(error)
-          }
-        }
-      }
-    }
-  }
-  
-  @objc func handleDebugAction(_ action: DebugAction, completion: @escaping RCTResponseSenderBlock) {
-    switch action {
-    case .fetchDiagnosisKeys:
-      manager.getDiagnosisKeys { (keys, error) in
-        if let error = error {
-          completion([error, NSNull()])
-        } else {
-          completion([NSNull(), keys!.map { $0.asDictionary }])
-        }
-      }
-    case .detectExposuresNow:
-      detectExposures { success in
-        if success {
-          completion([NSNull(), "Exposure detection successfully executed."])
-        } else {
-          completion(["Exposure detection error.", NSNull()])
-        }
-      }
-    case .simulateExposureDetectionError:
-      BTSecureStorage.shared.exposureDetectionErrorLocalizedDescription = "Unable to connect to server."
-      completion([NSNull(), "Exposure deteaction error message: \(BTSecureStorage.shared.exposureDetectionErrorLocalizedDescription)"])
-    case .simulateExposure:
-      let exposure = Exposure(id: UUID().uuidString,
-                              date: Date().posixRepresentation - Int(TimeInterval.random(in: 0...13)) * 24 * 60 * 60 * 1000,
-                              duration: TimeInterval(Int.random(in: 1...10) * 60 * 5 * 1000),
-                              totalRiskScore: .random(in: 1...8),
-                              transmissionRiskLevel: .random(in: 0...7))
-      let exposures = BTSecureStorage.shared.exposures
-      exposures.append(exposure)
-      BTSecureStorage.shared.exposures = exposures
-      completion([NSNull(), "Exposures: \(BTSecureStorage.shared.exposures)"])
-    case .resetExposureDetectionError:
-      BTSecureStorage.shared.exposureDetectionErrorLocalizedDescription = .default
-      completion([NSNull(), "Exposure Detection Error: "])
-    case .getAndPostDiagnosisKeys:
-      getAndPostTestDiagnosisKeys { error in
-        if let error = error {
-          completion([error.localizedDescription, NSNull()])
-        } else {
-          completion([NSNull(), "Local diagnosis keys successfully fetched and posted."])
-        }
-      }
-    case .resetExposures:
-      BTSecureStorage.shared.exposures = List<Exposure>()
-      completion([NSNull(), "Exposures: \(BTSecureStorage.shared.exposures.count)"])
-    case .toggleENAuthorization:
-      let enabled = manager.exposureNotificationEnabled ? false : true
-      requestExposureNotificationAuthorization(enabled: enabled) { result in
-        self.notifyUserBlueToothOffIfNeeded()
-        completion([NSNull(), "EN Enabled: \(self.manager.exposureNotificationEnabled)"])
-      }
-    }
-  }
-  
-  @objc func registerBackgroundTask() {
-    notifyUserBlueToothOffIfNeeded()
-    BGTaskScheduler.shared.register(forTaskWithIdentifier: ExposureManager.backgroundTaskIdentifier, using: .main) { [weak self] task in
-      
-      // Notify the user if bluetooth is off
-      self?.notifyUserBlueToothOffIfNeeded()
-      
-      // Perform the exposure detection
-      let progress = ExposureManager.shared.detectExposures { success in
-        task.setTaskCompleted(success: success)
-      }
-      
-      // Handle running out of time
-      task.expirationHandler = {
-        progress.cancel()
-        BTSecureStorage.shared.exposureDetectionErrorLocalizedDescription = NSLocalizedString("BACKGROUND_TIMEOUT", comment: "Error")
-      }
-      
-      // Schedule the next background task
-      self?.scheduleBackgroundTaskIfNeeded()
-    }
-  }
-  
+
 }
 
 private extension ExposureManager {
-  
+
   @objc func scheduleBackgroundTaskIfNeeded() {
     guard ENManager.authorizationStatus == .authorized else { return }
     let taskRequest = BGProcessingTaskRequest(identifier: ExposureManager.backgroundTaskIdentifier)
@@ -300,9 +231,11 @@ private extension ExposureManager {
       print("Unable to schedule background task: \(error)")
     }
   }
-  
+
   func notifyUserBlueToothOffIfNeeded() {
     let identifier = String.bluetoothNotificationIdentifier
+
+    // Bluetooth must be enabled in order for the device to exchange keys with other devices
     if ENManager.authorizationStatus == .authorized && manager.exposureNotificationStatus == .bluetoothOff {
       let content = UNMutableNotificationContent()
       content.title = String.bluetoothNotificationTitle.localized
