+++ ios/BT/Extensions/Foundation/Encodable+Extensions.swift
