+++ ios/BT/Extensions/Foundation/String+Extensions.swift
